lesson15 kadai9 first_submisson

+++ app/Http/Controllers/TasksController.php
@@ -17,13 +17,25 @@ class TasksController extends Controller
     // getでtasks/にアクセスされた場合の「一覧表示処理」
     public function index()
     {
-        // task一覧を取得
-        $tasks = Task::all();
+        // タスク一覧をidの降順で取得
+        //$tasks = Task::orderBy('id', 'desc')->paginate(10);
 
+
+        $data = [];
+        if (\Auth::check()) { // 認証済みの場合
+            // 認証済みユーザを取得
+            $user = \Auth::user();
+            // ユーザのタスク一覧を作成日時の降順で取得
+            $tasks = $user->tasks()->orderBy('created_at', 'desc')->paginate(10);
+
+            $data = [
+                'user' => $user,
+                'tasks' => $tasks,
+            ];
+        }
+        
         // task一覧ビューでそれを表示
-        return view('tasks.index', [
-            'tasks' => $tasks,
-        ]);
+        return view('tasks.index', $data);
     }
 
 
@@ -40,7 +52,7 @@ class TasksController extends Controller
         $task = new Task;
 
         // タスク作成ビューを表示
-        return view('tasks.create', [
+       return view('tasks.create', [
             'task' => $task,
         ]);
     }
@@ -65,11 +77,21 @@ class TasksController extends Controller
         ]);
         
         // taskを作成
-        $task = new Task;
-        $task->status = $request->status;    // 追加
-        $task->content = $request->content;
-        $task->save();
+        //$task = new Task;
+        //$task->status = $request->status;    // 追加
+        //$task->content = $request->content;
+        //$task->save();
 
+        
+        // 認証済みユーザ（閲覧者）の投稿として作成（リクエストされた値をもとに作成）
+        $request->user()->tasks()->create([
+            'content' => $request->content,
+            'status' => $request->status,
+        ]);
+
+        // 前のURLへリダイレクトさせる
+        //return back();　前のURL＝tasks.create(タスク作成ページ)なので何も動作が内容に見えた。２手間なので/にリダイレクトする。
+        
         // トップページへリダイレクトさせる
         return redirect('/');
     }
@@ -88,11 +110,20 @@ class TasksController extends Controller
     {
         // idの値でtaskを検索して取得
         $task = Task::findOrFail($id);
-
-        // task詳細ビューでそれを表示
-        return view('tasks.show', [
-            'task' => $task,
-        ]);
+            
+                    
+        // 認証済みユーザ（閲覧者）がその投稿の所有者である場合は、投稿を削除
+        if (\Auth::id() === $task->user_id) {
+            
+            // task詳細ビューでそれを表示
+            return view('tasks.show', [
+                'task' => $task,
+            ]);
+            
+        }
+        
+        // トップページへリダイレクトさせる
+        return redirect('/');
     }
 
 
@@ -111,9 +142,10 @@ class TasksController extends Controller
         $task = Task::findOrFail($id);
 
         // タスク編集ビューでそれを表示
-        return view('tasks.edit', [
-            'task' => $task,
-        ]);
+            return view('tasks.edit', [
+                'task' => $task,
+            ]);
+        
     }
 
 
@@ -161,8 +193,18 @@ class TasksController extends Controller
         // idの値でtaskを検索して取得
         $task = Task::findOrFail($id);
         // taskを削除
-        $task->delete();
+        //$task->delete();
 
+        
+
+        // 認証済みユーザ（閲覧者）がその投稿の所有者である場合は、投稿を削除
+        if (\Auth::id() === $task->user_id) {
+            $task->delete();
+        }
+
+        // 前のURLへリダイレクトさせる
+        //return back();　前のURL＝削除したタスク詳細ページなのでページが見つからずエラーになる。
+        
         // トップページへリダイレクトさせる
         return redirect('/');
     }
